Renamed context folder to stores folder
