+++ app/frontend/src/components/Disclaimer.tsx
@@ -11,7 +11,7 @@ import { useTranslation } from "react-i18next";
 import Cookies from "js-cookie";
 import { useMsal } from "@azure/msal-react";
 import { InteractionStatus } from "@azure/msal-browser";
-import { useAppStore } from "../context/AppStore";
+import { useAppStore } from "../stores/AppStore";
 
 export const Disclaimer = () => {
   const { t } = useTranslation();
