+++ app/frontend/src/components/TopMenu/subcomponents/ProfileMenuButton/index.tsx
@@ -3,10 +3,10 @@ import Menu from '@mui/material/Menu';
 import { Box, Chip, Collapse, Divider, FormControlLabel, FormGroup, FormLabel, ListItem, ListItemButton, ListItemIcon, ListItemText, Radio, RadioGroup, Switch, Typography } from '@mui/material';
 import { UserProfilePicture } from './ProfilePicture';
 import { useContext } from 'react';
-import { UserContext } from '../../../../context/UserContext';
+import { UserContext } from '../../../../stores/UserContext';
 import { useTranslation } from "react-i18next";
 import LanguageIcon from "@mui/icons-material/Language";
-import { useAppStore } from '../../../../context/AppStore';
+import { useAppStore } from '../../../../stores/AppStore';
 import HandymanIcon from '@mui/icons-material/Handyman';
 import PsychologyIcon from "@mui/icons-material/Psychology";
 import { allowedToolsSet, allowedCorporateFunctionsSet } from "../../../../allowedTools";
