+++ app/frontend/src/components/TopMenu/subcomponents/TopMenuFrame.tsx
@@ -6,7 +6,7 @@ import logo from "../../../assets/SSC-Logo-Purple-Leaf-300x300.png";
 import { type PropsWithChildren } from "react";
 import { type FC } from "react";
 import { AppBar } from "../../layouts/NewLayout/components/AppBar";
-import { useAppStore } from "../../../context/AppStore";
+import { useAppStore } from "../../../stores/AppStore";
 import { LEFT_MENU_WIDTH } from "../../../constants/frameDimensions";
 
 type TopMenuProps = PropsWithChildren<{
