+++ app/frontend/src/components/layouts/NewLayout/index.tsx
@@ -8,7 +8,7 @@ import MenuIcon from '@mui/icons-material/Menu';
 import { AppBar } from './components/AppBar';
 import { DrawerHeader } from './components/DrawerHeader';
 import { MainContentSection } from './components/MainContentSection';
-import { useAppStore } from '../../../context/AppStore';
+import { useAppStore } from '../../../stores/AppStore';
 import { LEFT_MENU_WIDTH } from '../../../constants/frameDimensions';
